+++ app/javascript/javascripts/application.js
@@ -1,3 +1,4 @@
+window.$ = require('jquery');
 $('#user_image').bind('change', function() {
   var size_in_megabytes = this.files[0].size/1024/1024;
   if (size_in_megabytes > 5) {
