--- db/migrate/20190904170449_add_provider_to_users.rb
@@ -1,7 +1,0 @@
-class AddProviderToUsers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :users, :provider, :string
-    add_column :users, :uid, :string
-    add_column :users, :username, :string
-  end
-end
