--- db/migrate/20190907190250_add_profile_image_to_users.rb
@@ -1,6 +1,0 @@
-class AddProfileImageToUsers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :users, :profile, :string
-    add_column :users, :image, :string
-  end
-end
