--- db/migrate/20190908193647_add_admin_to_users.rb
@@ -1,5 +1,0 @@
-class AddAdminToUsers < ActiveRecord::Migration[5.2]
-  def change
-    add_column :users, :admin, :boolean, default: false
-  end
-end
